Implement Close functionality and improve PlayPause responsiveness
Added a Close method to AudioEqualizerViewModel and updated the corresponding view to change "Cancel" to "Close." Enhanced PlayPause method in ShellViewModel for asynchronous execution and made various UI adjustments, including the addition of shuffle and repeat buttons and improvements to margin and padding in ShellView.xaml. Updated the installation script to increment the build version.